Conversion to Android studio project

+++ app/src/main/java/com/zach/tools/Reference/Parents/Children/Delivery.java
@@ -9,7 +9,7 @@ public class Delivery {
 	public Delivery(ContentValues cv){
 		data=cv;
 		if(data.containsKey("id")){
-			this.id = data.get("id");
+			this.id = (Integer)data.get("id");
 			data.remove("id");
 		}
 	}
